Reorganize code
- Move  into own submodule
- Rename instructions_8bitload etc -> execute_8bitload
- Update go.mod
- Combine some files
