+++ cpu/execute_16bitarith.go
