+++ cpu/execute_16bitload_test.go
