+++ cpu/execute_8bitarith.go
