+++ cpu/execute_8bitarith_test.go
