+++ cpu/execute_8bitload.go
