+++ cpu/execute_8bitload_test.go
