+++ cpu/execute_bitops.go
