+++ cpu/execute_control.go
