+++ cpu/execute_control_test.go
