+++ cpu/opcodes.go
