+++ go.mod
@@ -1,5 +1,3 @@
-module github.com/mpingram/gameboy-emu/m
+module github.com/mpingram/gameboy-emu
 
 go 1.13
-
-require github.com/mpingram/gameboy-emu v0.0.0-20190524020622-49fba869781e
