+++ main.go
@@ -1,0 +1,8 @@
+package main
+
+import (
+	_ "github.com/mpingram/gameboy-emu/cpu"
+)
+
+func main() {
+}
